Backup/Dashboard_removed_2025-12-12/Dashboard/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pandas==2.2.3
gitpython==3.1.43
pydantic==2.9.2
python-multipart==0.0.12
